requirements.txt
torch
torchvision
opencv-python
pillow
matplotlib
transformers
accelerate
timm
safetensors
sentencepiece
pyyaml\n